HYD-7179 Failure in upgrade process does not end the upgrade test
It would seem that we are not propagating errors in upgrade_os() up to
it's caller and then not killing the test in the overall upgrade_os
test if the upgrade_os() function does have a failure in it.

This was noticed when upgrade_os() was trying to remove the O/S provided
yum repo files and they don't exist on RHEL yet this was not causing the
error to propagate up.  So might as well fix that also.

 - add error propagation to upgrade_os()
 - fail the upgrade_os test if upgrade_os() reports an error

Reviewed-on: https://review.whamcloud.com/25696
Tested-by: Jenkins
Tested-by: Chroma Test User
(cherry picked from commit 3760a269c0f8ec9ba7ec92168a6a7e589232a3cd)
Reviewed-on: https://review.whamcloud.com/26179

+++ chroma-manager/tests/framework/integration/utils/upgrade_os.sh
@@ -61,12 +61,18 @@ fi
 
 rpm -qa | sort >/tmp/after_upgrade
 
-# The yum upgrade will have restored the /etc/yum.repos.d/CentOS-*.repo
+# The yum upgrade might have restored the /etc/yum.repos.d/CentOS-*.repo
 # files so remove them here
-rm /etc/yum.repos.d/CentOS-*.repo
+if ls /etc/yum.repos.d/CentOS-*.repo; then
+    rm /etc/yum.repos.d/CentOS-*.repo
+fi
 
 # TODO: we really ought to reboot here
 #       a new kernel was surely installed and real users would reboot here" | dshbak -c
+
+    if [ ${PIPESTATUS[0]} != 0 ]; then
+        return 1
+    fi
 
     echo "End operating system upgrade"
